Art gallery first try
idea: I thought about the page being responsive, and for many images,
the page should be able to scroll. so I thought of repeating a single
pixel line of the bg image. This doesn't work very well. my second idea
will be to wrap the images in a overflow: scroll box